start stat library and move track code out of analysis.hh
